Remove unnecessary casts.

+++ src/test/java/org/apache/commons/io/CopyUtilsTest.java
@@ -62,7 +62,7 @@ public class CopyUtilsTest {
     private static final int FILE_SIZE = 1024 * 4 + 1;
 
 
-    private final byte[] inData = TestUtils.generateTestData((long) FILE_SIZE);
+    private final byte[] inData = TestUtils.generateTestData(FILE_SIZE);
 
     // ----------------------------------------------------------------
     // Tests
